+++ changelog.md
@@ -21,6 +21,8 @@
    - Added `visibility` attribute to **GET** `/economy/balance`
    - Included all types in transaction list & export
    - Added the forgotten `wear_ratio` calculation
+   - Limited job income to `4294967296`
+   - Added **PATCH** `/balance/{userid}` to update balance without making transaction
 
 ## v2.8.4